Qt6: Ensure QT_STYLE_OVERRIDE is cleared.
Some distros may set QT_STYLE_OVERRIDE by default when attempting to
style their Qt5 applications. However, this style may not necessarily
exist for Qt6, which will result in a failure to load QML content. To
work around this issue, ensure that the styling is disabled before
creating the QApplication.

The QML fails to load QtQuick.Controls when this situation occurs.